Move imageio functionality into IO module
