--- brainglobe_utils/image_io/__init__.py
@@ -1,6 +1,0 @@
-__author__ = "Charly Rousseau, Adam Tyson"
-__version__ = "0.2.4"
-
-from brainglobe_utils.image_io.load import *
-from brainglobe_utils.image_io.save import *
-from brainglobe_utils.image_io.utils import *
